Send assets to browser
The Head component is not updated whenever assets are added.
Might have to make a VHead class...

Or maybe better to not let the user create the head tags.
Maybe the user should only be able to create the body tag,
so that head will be 100% controlled by the vdom runtime.